# docs/specs/ecardo-travel-openapi.yaml
openapi: 3.1.0
info:
  title: eCardo Travel API
  version: 1.0.0
  description: Stable Flutter and website API for eCardo Travel. Requests enter through the Cloudflare edge, provider payloads are normalized before reaching the private origin, and all purchases use the authenticated user's eCardo wallet.
servers:
  - url: https://trip.ecardo.ir/api
...
security:
  - bearerAuth: []
paths:
  /v1/travel/bootstrap:
    get:
      tags: [Discovery]
      security: []
      summary: Get enabled services and server-driven UI schemas
      parameters:
        - { $ref: '#/components/parameters/Locale' }
      responses:
        '200':
          description: Service catalog
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BootstrapResponse' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: eCardo-Travel-Token }
  parameters:
...
    Locale: { name: locale, in: query, schema: { type: string, example: en-US }, description: Locale values are normalized case-insensitively to a supported two-letter language code. X-Locale is also accepted. }
...
  schemas:
    BootstrapResponse:
      type: object
      properties:
        status: { type: string, const: success }
        schema_version: { type: string, const: '1.0' }
        data:
          type: object
          properties:
            brand: { type: string }
            currency: { type: string }
            locale: { type: string }
            services: { type: array, items: { type: object, additionalProperties: true } }
...